Improves visual appeal with background images
Adds background images to the layout and cycles them on route changes to improve the visual appeal.

Introduces a background with a subtle radial gradient and adds a subtle transition for a smoother user experience.

Also, moves the quote block from the navigation to the home page and enhances its appearance.