Adds logo and restructures navigation
Improves the site's visual identity by adding a logo to the navigation bar.

Enhances the navigation structure by removing the direct link to the home page from the main navigation data and instead linking to the root via the logo. This declutters the navigation and provides a more intuitive user experience.